Moved Callback class in remote package
