+++ CaughtUp/app/src/main/java/news/caughtup/caughtup/ws/remote/Callback.java
@@ -1,4 +1,6 @@
-package news.caughtup.caughtup.entities;
+package news.caughtup.caughtup.ws.remote;
+
+import news.caughtup.caughtup.entities.ResponseObject;
 
 public abstract class Callback {
     public abstract void process(ResponseObject responseObject);
