+++ CaughtUp/app/src/main/java/news/caughtup/caughtup/ws/remote/IRest.java
@@ -1,7 +1,5 @@
 package news.caughtup.caughtup.ws.remote;
 
-import news.caughtup.caughtup.entities.Callback;
-
 public interface IRest {
     void getCall(String endPoint, String jsonData, Callback callback);
     void postCall(String endPoint, String jsonData, Callback callback);
